- Step2 코드를 조금 더 효율적으로 작성하기 위해서 수정단계     1. Game이라는 클래스를 다른 하위클래스(?)가 입력받는 형태가        이상하다고 생각되어서 GameInfo라는 정보를 담아두는 클래스를 하나        생성하고, 해당 클래스에 게임을 진행하는데 필요한 정보를        저장하고자 하였다.     2. 메소드의 길이를 10줄이 넘지 않게 하기 위해서 줄일 수 있는        부분들을 수정하였다.
